Product Model and InMemory Product Repository Built
