+++ MyShop/MyShop.WebUI.Tests/packages.config
@@ -4,7 +4,7 @@
   <package id="Microsoft.AspNet.Razor" version="3.2.7" targetFramework="net472" />
   <package id="Microsoft.AspNet.WebPages" version="3.2.7" targetFramework="net472" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net472" />
-  <package id="MSTest.TestAdapter" version="1.2.0" targetFramework="net472" />
-  <package id="MSTest.TestFramework" version="1.2.0" targetFramework="net472" />
-  <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net472" />
+  <package id="MSTest.TestAdapter" version="2.2.6" targetFramework="net472" />
+  <package id="MSTest.TestFramework" version="2.2.6" targetFramework="net472" />
+  <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net472" />
 </packages>
