+++ MyShop/MyShop.WebUI/Scripts/jquery-3.6.0.intellisense.js
