+++ MyShop/MyShop.WebUI/Scripts/jquery.validate-vsdoc.js
@@ -4,7 +4,7 @@
 * intended to be used only for design-time IntelliSense.  Please use the
 * standard jQuery library for all production use.
 *
-* Comment version: 1.17.0
+* Comment version: 1.19.3
 */
 
 /*
@@ -15,7 +15,7 @@
 * for informational purposes only and are not the license terms under
 * which Microsoft distributed this file.
 *
-* jQuery Validation Plugin - v1.17.0 - 12/5/2016
+* jQuery Validation Plugin - v1.19.3 - 8/12/2020
 * https://github.com/jzaefferer/jquery-validation
 * Copyright (c) 2013 Jörn Zaefferer; Licensed MIT
 *
